fix class attribute dispatching code - ignore attributes of `abc.ABC` class
otherwise isinstance/issubclass might give incorrect results due to
`ABC._abc_cache` confision. ClassMappingDispatcher is a subclass of
ABCMeta, and therefore all the classes that use ClassMappingDispatcher
as a metaclass will have the attributes and functionality of `abc.ABC`.
To avoid confusion within this functionality, we should not interfere
with this attributes.