revert the decision to convert Decimal -> float in RPC parsed results
This might break code, and such change is unnecessary.
COIN value for Bitcoin is selected in such a way that there would be
no rounding losses when rounded to 8th fractional digit, so converting
to float might be OK, but this is really a breaking change, for no
additional benefit. We allow Decimal as argument to
`coins_to_satoshi()`, instead.