fix regression on pypy3 with ctypes
`CPubKey._to_raw()` should return ctypes char-array, but
was erroneously made to return bytes. This caused regression on
pypy3 (on python3 everything was fine).
This was a result of confusion over the name of the method
and the fact that ctypes char buffer has 'raw' attribute
that is a bytes instance.

To prevent this in the future, `_to_raw()` is renamed to
`_to_ctypes_char_array()`, and `_from_raw()` is renamed to
`_from_ctypes_char_array()`